パッケージを構造の改善とpackage-info.java の加筆・修正（意図を明確に）
