+++ src/main/java/example/application/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * アプリケーション層
+ * 実現するアプリケーション機能を表現する
+ * ドメイン層へのファサード
+ */
+package example.application;
