+++ src/main/java/example/application/service/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * サービス定義
+ * 個々の機能部品を定義する
+ */
+package example.application.service;
