+++ src/main/java/example/application/usecase/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * ユースケース定義
+ * 機能（サービス）を組み合わせた、ユースケースシナリオを記述する
+ */
+package example.application.usecase;
