--- src/main/java/example/datasource/package-info.java
@@ -1,4 +1,0 @@
-/**
- * DataAccess/Infrastructure層です。
- */
-package example.datasource;
