+++ src/main/java/example/domain/fundamentals/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * 基礎部品
+ * ドメインモデルを表現するための基礎部品を記述する
+ * 基本的な値オブジェクトなど
+ */
+package example.domain.fundamentals;
