+++ src/main/java/example/domain/model/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * 中核のモデル
+ * fundamentals や Java の基本ライブラリで、ドメインの重要な関心事を表現する
+ */
+package example.domain.model;
