+++ src/main/java/example/domain/model/user/GenderType.java
@@ -1,0 +1,9 @@
+package example.domain.model.user;
+
+public enum GenderType {
+
+    男性,
+    女性,
+    不明
+
+}
