+++ src/main/java/example/domain/model/user/UserRepository.java
@@ -1,0 +1,19 @@
+package example.domain.model.user;
+
+/**
+ * Created by haljik on 15/06/04.
+ */
+public interface UserRepository {
+    User findBy(UserIdentifier id);
+    Boolean isExist(UserIdentifier id);
+
+    UserSummaries list();
+
+    User prototype();
+
+    void register(User user);
+
+    void update(User user);
+
+    void delete(User user);
+}
