+++ src/main/java/example/domain/model/user/UserSummary.java
@@ -1,0 +1,14 @@
+package example.domain.model.user;
+
+public class UserSummary {
+    UserIdentifier identifier;
+    Name name;
+
+    public UserIdentifier identifier() {
+        return identifier;
+    }
+
+    public Name name() {
+        return name;
+    }
+}
