+++ src/main/java/example/domain/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * ドメイン層
+ * 業務の知識、利用者の関心事を記述する
+ */
+package example.domain;
