+++ src/main/java/example/infrastructure/datasource/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * データベースアクセスの実装クラス群
+ */
+package example.infrastructure.datasource;
