+++ src/main/java/example/infrastructure/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * 技術基盤
+ *
+ * このレイヤのクラスを他のレイヤのクラスは直接呼び出さないこと
+ * 使う側の意図をインタフェースで宣言
+ * 実装クラスを、このレイヤで宣言（命名ガイド：実装技術名＋インタフェース名）
+ *
+ */
+package example.infrastructure;
