+++ src/main/java/example/infrastructure/transfer/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * トランスファー層
+ *
+ * 外部との通信の実装クラス
+ */
+package example.infrastructure.transfer;
