--- src/main/java/example/model/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Domain Model です。
- */
-package example.model;
