--- src/main/java/example/model/user/GenderType.java
@@ -1,9 +1,0 @@
-package example.model.user;
-
-public enum GenderType {
-
-    男性,
-    女性,
-    不明
-
-}
