--- src/main/java/example/model/user/UserSummary.java
@@ -1,14 +1,0 @@
-package example.model.user;
-
-public class UserSummary {
-    UserIdentifier identifier;
-    Name name;
-
-    public UserIdentifier identifier() {
-        return identifier;
-    }
-
-    public Name name() {
-        return name;
-    }
-}
