+++ src/main/java/example/presentation/controller/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * コントローラ定義
+ */
+package example.presentation.controller;
