+++ src/main/java/example/presentation/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * プレゼンテーション層
+ * ビューとコントローラの置き場所
+ */
+package example.presentation;
