+++ src/main/java/example/presentation/view/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * ビュー定義
+ * ビューを表現するクラス群
+ *
+ * JSON <-> オブジェクト マッピングクラス
+ * View Helper
+ * etc.
+ */
+package example.presentation.view;
