--- src/main/java/example/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Service層 ユースケースとなるサービス郡です。
- */
-package example.service;
