--- src/main/java/example/web/package-info.java
@@ -1,4 +1,0 @@
-/**
- * web層 ControllerなどWebのUIに関わるオブジェクト郡です。
- */
-package example.web;
