+++ src/main/resources/application.properties
@@ -3,6 +3,6 @@ spring.thymeleaf.cache=false
 
 spring.datasource.sql-script-encoding=utf-8
 
-logging.path=.
+#logging.path=/tmp
 
 mybatis.config-location=classpath:mybatis.xml
