+++ src/test/groovy/example/application/service/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * ユースケースの自動テスト郡です。
+ */
+package example.application.service;
