--- src/test/groovy/example/service/package-info.java
@@ -1,4 +1,0 @@
-/**
- * ユースケースの自動テスト郡です。
- */
-package example.service;
